Multi

+++ release/release.bash
@@ -11,6 +11,15 @@ mkdir -p $BIN_DIR
 echo "Creating ${RELEASE_NAME}.tar.gz..." 1>&2
 "$SRC/release/build.bash" "$BIN_DIR/$BINARY" "$RELEASE_VERSION"
 tar -C "${BIN_DIR}" -cvzf "${RELEASE_NAME}.tar.gz" "${BINARY}"
-docker run --rm -v ${BIN_DIR}:/app xitonix/fpm-rpm -s tar -t rpm -n ${BINARY} -p /app -v ${RELEASE_VERSION} --iteration ${RPM_ITERATION} "/app/${RELEASE_NAME}.tar.gz"
+echo $(pwd)
+ls -al $(pwd)
+if [ ! -z "$1" ]
+then
+  docker run --rm -v ${BIN_DIR}:/app xitonix/fpm-rpm -s tar -t rpm -n ${BINARY} -p /app -v ${RELEASE_VERSION} --iteration ${RPM_ITERATION} "/app/${RELEASE_NAME}.tar.gz"
+fi
 echo "::set-output name=file::${RELEASE_NAME}.tar.gz"
-echo "::set-output name=rpm::${BINARY}-${RELEASE_VERSION}-${RPM_ITERATION}.x86_64.rpm
+
+if [ ! -z "$1" ]
+then
+  echo "::set-output name=rpm::${BINARY}-${RELEASE_VERSION}-${RPM_ITERATION}.x86_64.rpm
+fi
